requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=2.8.1
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1

# Redis for conversation storage
redis>=5.0.0

# Pinecone vector database
pinecone>=5.0.0

# Tavily web search (supplementary)
tavily-python>=0.5.0

# PostgreSQL for hybrid RAG queries
psycopg2-binary>=2.9.9

# PDF and document processing
PyMuPDF==1.24.0
Pillow==10.2.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
tabulate==0.9.0
